Fix function context register handling.
We cannot just use a new virtual register for it, since the assignment might be
in a branch that is not necessarily executed.

It's pretty weird that anything worked even though this bug was present! But
apparently this resulted in pretty wide live ranges and a lot of registers got
spilled... dunno.

Also adds some debug prints.