Register allocator fix.
The bug is always in the register allocator.

It was allocating virtual registers to real registers which were marked
as conflicting. This happened because the real register was
"simplified" (i.e., detected to have so few conflicts that we can always
find a real register for it), and simplified nodes were not taken into
account in the conflict detection (actually, why??).

Solution: Don't simplify real registers, it doesn't make sense.